min max input challenge

+++ src/InputChallenge.java
@@ -5,7 +5,7 @@ public class InputChallenge {
         Scanner scanner= new Scanner(System.in);
 
         int counter=1;
-        int sum=0;
+        double sum=0d;
 
 
 
@@ -13,7 +13,7 @@ public class InputChallenge {
             System.out.println("Enter number #" + counter + ": ");
             String nextNumber= scanner.nextLine();
             try{
-                int number = Integer.parseInt(nextNumber);
+                double number = Double.parseDouble(nextNumber);
                 counter++;
                 sum += number;
             }
